moved camera to another folder
